+++ src/main/java/TPZTBCS/Prenda.java
@@ -27,7 +27,7 @@ public class Prenda
 	@Id
     @GeneratedValue(strategy = GenerationType.AUTO)
 	@Column(name = "ID_PRENDA")
-	Long ID;
+	int ID;
 	
 	@ManyToOne
 	@JoinColumn (name = "atuendo_id")
